* Submitted by: Tim Prins  * Reviewed by: Josh Hursey
Fixes #440
Refs #369

LoadLeveler? only sets LOADL_PROCESSOR_LIST when there are 128 or less
tasks allocated to a job. The POE RAS relied on this variable so I
created a new RAS which uses the LoadLeveler? API instead of relying
on the environment variable. This still needs some testing, so for now
we use the POE RAS whenever LOADL_PROCESSOR_LIST, otherwise we fall
back on this component. 

Unfortunately, this will require an autogen... 

-----

 * Formatting and minor cleanup 
 * made it so we now set the architecture of each node we discover 
 * remove debugging output 

-----

cleanup