* Submitted by: Brian Barrett  * Reviewed by: Ralph Castain, Jeff Squyres
Fixes #460

num_procs can be zero if something went partly wrong before. This will 
cause a math exception on some platforms, so don't let that happen.

Had to ammend the original r11929 because it includes a Windows fix
that did not exist on the v1.1 branch -- this caused a trivial
conflict that was easy to solve (and Windows isn't supported on the
v1.1 branch, so the change is irrelevant).