Ticket #417: openib subnet fix
Local GID table contains not what I thought it contains. It contains local HCA 
GIDs (there can be more than one) and not GIDs of the HCA on the network. Entry
zero always have to be initialized so we use it, and warn user if there is more 
then one port active and default subnet is configured on at least one of them.

Merged r11815, which closes #417.

Submitted by: gleb, Reviewed by: Jeff and Roland Drier, Merged by: Tim Mattox