* Submitted by: Jeff Squyres  * Reviewed by: Brian Barrett
Fixes ##527.

From r12246 on the trunk:

Fix a problem when $FC/$F77 were multiple tokens. 

In addition to the r12246 patch, massaged the NEWS patch to ''not''
include the bit about MPI_GROUP_TRANSLATE_RANKS because that has not
yet come over to the v1.1 branch.  Also removed a chunk from the
ompi_setup_f90.m4 patch because it isn't relevant on the v1.1 branch
(where we don't allow non-static building of F90 libraries).