Ticket #536: pml pipeline and self descriptor managing fixes
Merged r12291 and r12292 which closes #536, refs #246.

r12291: (reviewed by George)
Fix a bug in descriptor handling code. The self BTL was mixing the different 
kinds of descriptors (e.g. put rdma descriptor in the eager free-list). 

r12292: (reviewed by gleb)
Increment the pipeline depth before we trigger the send function. As 
mentioned in the comment the completion/callback of the triggered 
send operation can happen before the call returns. If this happens and 
if the pipeline depth is 0 before we triggered the send operation and 
this is the last send operation of the request then the completion detection 
code will decrement the pipeline depth and check it for equality to 0. 
Because (0-1) != 0 the pml completion function for this request will 
*not* be called.

Submitted by hpcstork, Reviewed by George & gleb, Merged by Tim Mattox