Ticket #572: ob1 fixes for MPI_*send_init_* Intel tests
Merged r12342 which closes #572, refs #547.

r12342:
Do not modify the convertor if there is no data to be send across the network.
The req_bytes_packed field is initialized in the BASE_INIT macro, so it is
set for all requests at this stage. 

Submitted by tprins, Reviewed & Merged by Tim Mattox