Ticket #596: OS X PPC64 compile failure fix
Merged r12418 which closes #596, refs #575.

r12418:
Run configure check for the registers in the thread state to
decide whether they are prefixed with a double underscore or not.
DARWIN_UNIX03 is defined on Tiger when compiling PPC 64 code, so
we can't use that.

Submitted by Brian, Reviewed by Jeff Squyres, Merged by Tim Mattox