Ticket #592: Use a new basic algorithm for allgatherv.
Merged r12469, r12493 which closes #592.

r12469:
Use a new algorithm for allgatherv.
The old algorithm essentially did N gatherv's:

  for (i = 0 ... size)

  MPI_Gatherv(..., root = i, ...)

The new algorithm simply does (effectively):

  MPI_Gatherv(..., root = 0, ...)
  MPI_Bcast(..., root = 0, ...)

r12493:
Remove the call to set_args. This is only required for the MPI level,
because there we have to be able to return to the user the description
of the data.

Submitted by jsquyres, Reviewed by bosilca, Merged by Tim Mattox