* Submitted by: Jeff Squyres  * Reviewed by: Brian Barrett
Fixes #601

This is a workaround to bug in the Intel C++ compiler, version 9.1 
(all versions up to and including 20060925). The issue has been 
reported to Intel, along with a small [non-MPI] test program that 
reproduces the problem (the test program and the OMPI C++ bindings 
work fine with Intel C++ 9.0 and many other C++ compilers). 

See the lengthy commit message in r12514 for a full description of
what this commit does and why.