Ticket #599
Restore the ability to connect to default universes

This patch consists of two patches:

r12438 and r12472

Collectively, they restore the ability to connect to default universes in
support of persistent universe operations. These resolved several tickets that
were on the trunk (those tickets were closed out on the trunk) - this changeset
brings those repairs to the 1.2 branch.

Submitted by: Ralph Castain
Reviewed by: Rolf Vandeavaart