Closes #631: one-sided fixes
Submitted by brbarret, Reviewed by Galen, Merged by Tim Mattox

r12667:
 * Use one-sided datatype check instead of send/receive and check both
the origin and target datatypes 
 * Allow error handler to be set on MPI_WIN_NULL, per standard 
 * Allow recursive calls into the pt2pt osc component's progress function 
 * Fix an uninitialized variable problem in the unlock header 

r12673:
 * Fix a counter roll-over issue that could result from a large
(but not excessive) number of outstanding put/get/accumulate calls
during a single synchronization issues
 * Fix epoch issue with rdma component that would effect PWSC synchronization

r12675:
Remove check for epoch on the receiver side, as this is a race
condition between message arrival, and because we'll be still
blocking in a synchronization call, is of no consequence to the user.
We do an epoch check on the sending side, so this isn't an issue there either.

r12676:
Fix issue where messages could start arriving before the window was
fully initialized, especially during lock/unlock, which doesn't use MPI
collectives for synchronization (unlike Fence)