Closes #664: add_error_class() and add_error_code() fixes Closes #667: Update the xcast subsystem to avoid tcp stack overload
These two CMRs overlapped, so got merged together in one checkin.

Submitted by rhc & edgar, Reviewed by Jeff, Patched by Tim Mattox

== The xcast patch corresponds to parts of r12679, r12706, and r12722:

The previous xcast in v1.2 used non-blocking sends to transmit the
buffers to the remote application processes. This can overload tcp
stacks as the size of the application and/or the size of the buffer
grows.  This patch replaces the non-blocking sends with blocking sends.
It also provides an optional transmission mode based on a binomial tree.
Refs #641

== The add_error_class patch corresponds to parts of r12764:

Rewriting much of the errorcode and errorclass code, since 
 1. we have to be able to attach a string to an error class,
    not just to an error code 
 2. according to MPI-2 the attribute MPI_LASTUSEDCODE has to
    be updated everytime you add a new code or a new class.
    Thus, you have to have single list for both. 

Thus, we got rid of the error_class structure. In the error-code
structure, we can distinguish whether we are dealing with an error
code or an error class by looking at the err->code element of
the structure. In case its value is MPI_UNDEFINED, the according
entry is a class, else it is an error code. All predefined error
codes have the code and the class field set to the same value. 
The test MPI_Add_error_class1 passes now.
Refs #418