Closes #717: Repair singleton operations to remove inaccurate error message
Submitted by Ralph, Reviewed by Jeff, Merged by Tim Mattox

r12952:
Fix a buglet in the singleton startup procedure. For purposes
of minimizing the xcast message, we "strip" the descriptive
info on all subscription messages. This means, though, that we
have to store the process name and other info so it can be
retrieved in the body of the subscription data (as opposed to
in the description). This wasn't being done for singletons
because they don't call the RMAPS to "map" themselves. 

This has now been corrected. The singleton startup will dutifully
call the mapper framework so that the proper data storage
locations get initialized. Unfortunately, we then had to instruct
the RMAPS not to allocate a vpid range for this job - otherwise,
it would make a mistake and think there were two processes in it.
Hence, a change was required to RMAPS to tell it "map this job,
but don't allocate a vpid range for it".