Refs #796
One-sided fixed into 1.2

Send reply from the passive side of an unlock request back to the active
side and only let MPI_WIN_UNLOCK return when the passive side has actively
replied that the window is unlocked. 

Don't call ompi_request_test() if the request isn't likely to finish.
Otherwise, we end up recursively calling into the progress functions
and corrupting a list that doesn't like to be corrupted.

Fix an issue with recursive calls into the component progress caused by btls
sometimes calling opal_progress()
during their send calls by dropping the loop through the list of pending control
messages if any are marked
as completed. 

Send the unlock request before starting the requests. We won't unlock until we
get an ack from the remote side,
so there's no longer a race there (I used to do the unlock request last, after
local completion of all the
requests completed, to try to avoid having the passive side reply to the active
side, but I don't do that
anymore). The unlock side will not "unlock" the window until it actually
receives the correct number of results,
so we're good there.

This fixes an issue where we would receive data on the remote side we weren't
expecting that could cause
us to release a lock before it really should have been released to the
requesting peer. It could also
cause a deadlock if one of the processes trying to unlock was "self", as that
would result in the active
unlock never sending the unlock request, even though it sent the payload, which
could cause a counter
that should always be positive to hit -1, causing an infinite loop that could
only be solved by
popping up the stack, which was an impossibility. 

Submitted by: Brian
Reviewed by: Galen