Closes #799: buffer detach and buffered send alignment fix
Submitted by rolfv, Reviewed by jsquyres, Merged by Tim Mattox

r13205:
Fix two problems. For MPI_Buffer_detach, do not attempt to 
return the buffer address from Fortran. It is not expected 
behavior. For MPI_Buffer_attach, adjust the address of 
the buffer handed in so it is always aligned.