Closes #809: fixes --prefix behavior under SLURM
Submitted by Jeff, reviewed by Ralph Castain, Merged by Tim Mattox

r13239:
For --prefix-like behavior, we used to modifiy environ directly and 
then exec the "srun..." from there. But somewhere along the line, we 
switched to having a copy of environ and modifying that. It looks 
like we forgot to update the stuff for --prefix behavior. So this 
commit fixes the setenv's for PATH and LD_LIBRARY_PATH to modify the 
environ copy (not environ itself) so that the values properly get 
passed down to the srun environment via execve(). 

This restores --prefix behavior in the SLURM pls.