Closes #810: ORTE changes that update pernode/npernode and other features
Submitted by rhc, Reviewed by Jeff, Brian and George, Merged by Tim Mattox

Code drop from ORTE tmp branch:
This changeset "sweeps" up a number of things sitting in the
trunk, most of them truly minor, but a couple that are "fixes"
to significant customer-requested behavior. The minor changes
are being swept in to make the branch easier to manage - while
minor in nature, they hit a bunch of files and make it nearly
impossible to find the substance in the chaff.

The vast majority of lines in this change come from:
 1. update to a whole bunch of configure.params files - just
   remove a defunct line and change the copyright in each
   one (written by Brian, reviewed by me)
 2. add some orte tests (including unit tests) just to maintain
   track with the trunk (written by me - no review required as
   they don't impact the code)

The real substance is in the remaining few lines of code:
 1. fix the pernode and npernode behavior to incorporate limits
   on the total number of procs being spawned. We now check
   if -np is provided and use that as a limiter (Jeff should
   probably review since it is for his customer)
 2. put a check in the release of orte_process_name_t data types
   to ensure the pointer isn't null (a very rare error we were seeing)
   (written by me - about the most trivial change you will ever see).
 3. fix a few minor buglets that had been found in the code (can't
   recall all of them, but they were pretty minor things) (written
   by various people - reviewed by me)