Fix the MPI_ABORT problem.  We added an orte_initialized variable yesterday and set it to "true" in orte_init().  But ompi_mpi_init() doesn't call orte_init() -- it calls orte_init_stage1() and orte_init_stage2(). So orte_initialized was never set to true, and Badness happend from there (w.r.t. ompi_mpi_abort()).
This patch moves the setting of orte_initialized to orte_init_stage2()
so that everyone will always get it set properly.

It also moves setting orte_universe_info.state to RUNNING into
stage2() as well -- Ralph confirmed that that should have been there
for the same reasons that orte_initialized needs to be there.