Closes #847: rsh pls EFAULT (errno=14) fix
Submitted by George, Reviewed by Jeff, Merged by Tim Mattox

r13321:
We have this one starting from r12059. We save a pointer
to the argv[*] and then we modify the argv, forcing the
reallocation of the array. With luck the saved pointer
still have a meaning ... without execve return with
error 14 (EFAULT).