* Submitted by: Brian Barrett  * Reviewed by: George Bosilca
Fixes #907

r13644
use memory barriers for lock init and unlock 

r13649
A slightly more correct fix for issues with reordering unlock
calls. We want to flush all writes pending (ie, the data being
protected) out of the memory manager before we write the spinlock
unlock. Only need a wmb instead of full mb, which is at least slightly
less intrusive. Also, after much thought, no need for a memory barrier
in init.

r13651
Work around what appears to be an odd error in the pathscale compiler
where the dispatch function opal_atomic_cmpset_acq_xx was used to call
opal_atomic_cmpset_acq_32 would cause the compiler / linker to fix up
a jump address to be itself, leading to an infinite loop.

We're still looking into exactly what caused this, but during the
investigation into the hang, we determined that the compiler (both
pathcc and gcc) weren't always inlining both the call to
opal_atomic_cmpset_acq_xx and opal_atomic_cmpset_acq_32, meaning there
was a function call in opal_atomic_lock. The atomic lock will always
be 32 bit, so there's no need for the dispatch function, so might as
well remove the dispatch function that may or may not be inlined.

A bug fix that leads to potentially better performance. Gotta love the
few times that happens...