* Submitted by: Rich Graham  * Reviewed by: Jeff Squyres
Fixes #927

Port of r12949 and r12964 from the trunk to fix an sm btl startup race
condition.

There was a minor debate about whether to back-port this fix or not
(i.e., just release v1.1.5 with the problem and encourage anyone who
saw the problem to upgrade to v1.2).  Turns out that there are at
least a few people who are seeing this problem, and it was a trivial
back-port.  So it seemed like the Right thing to do.