Bump the RPM version number.
This is actually symbolic (and a slight mis-use of the RPM version
number) that OFED 1.2 has been given an RPM that includes all changes
on this branch up until this r number (the OFED 1.2 Open MPI v1.2ofed
tarball is actually marked as "r14110" in its VERSION file, but
nothing has happened on this branch in the OMPI code base since r14110
-- only some minor specfile changes).  

Hence, OFED v1.2 currently includes Open MPI v1.2ofed with RPM version
1 ("1.2ofed-1").

I am noting this here in the log because I am about to merge over all
the changes to the v1.2 branch into this branch in a first step
towards merging this branch back to the proper v1.2 branch.

As such, OFED 1.2 currently includes OMPI from this branch
'''BEFORE''' this commit.  If we decide to release OMPI v1.2ofed to
OFED 1.2 '''after''' this commit, we'll use the new RPM version
number, meaning that it'll be RPM version 1.2ofed-2.