Closes #968: Repair reuse-daemons
Submitted by rhc, Reviewed by tprins, RM-approved on the Tuesday telecon.

Applied patch:
An error patch to fix reuse-daemons operations failed to make
it over to the 1.2 branch in time for the 1.2 release (probably
my fault - didn't get filed). The error lies in the function
that gets the necessary data for the odls to launch the additional
procs. Basically, that function must *not* insert tokens into
the values being passed to the odls so that the launch_procs
function can correctly parse the launch data.

The attached patch removes the offending line, makes a corresponding
correction to the number of tokens in one other line, and inserts
a comment to ensure this doesn't recur.