Closes #627: Make TM components use $PBS_NODEFILE
Submitted by rhc, Reviewed by jsquyres, RM-Approved by Rich

r13581:
Bring in a forgotten speed improvement for the TM launcher that
was developed during SNL Tbird testing last year. Remove the
redundant and slow calls to TM to resolve hostnames. Instead,
read the host info from the PBS file during the RAS, and then
just use that info in the PLS (rather than getting it again). 

Adjust the RMAPS mapped_node object to propagate the required
launch_id info now included in the ras_node object. This provides
support for those few systems that don't use nodename to launch,
but instead want some id (typically an index into the array of
allocated nodes). This value gets set for each node in the RAS -
the RMAPS just propagates it for easy launch.

r14121:
Let's try to detect the default nodefile directory because it
can different for different sites. If we cannot detect the
default then we fall back to the hard coded path.