Add some comments on the internals of the bucket structure. Alter the cleanup function to make it more scalable. The memory fragmentation is still high, but at least in most of the cases (where all ressources are correctly released before the cleanup) the code is now highly efficient. Before the code execute in (N * (N-1))!, which take a while when the number of allocated ressources increase (which is the case when a lot of unexpected messages are created).
The fix consist of checking if all items are freed and if it's the case
then do not recreate the free items list (as we know that everything will
be released). If this condition is not true, we fall back on the
original execution path (which is still sub-sub-sub ... optimal).