Closes #994: ROMIO configure fix
Submitted by Jeff, Reviewed by Brian, RM-Approved by tdd

r14387:
A minor change to ROMIO's configure script:
make it use exactly the same "restrict" check as the top-level
OMPI configure.ac script so that it will guarantee to always
get the same result. Therefore, the #define for restrict will
always have the same value in both opal_config.h and romioconf.h,
and we get 7 less warnings (6 in the IO ROMIO component, 1 in
ROMIO itself) when compiling with icc on Linux (because
PAC_C_RESTRICT and AC_C_RESTRICT would get different values 
for the "restrict" #define in this case).